Create a Python package for WISE
This commit turns WISE into a Python package that declares its own
dependencies, can be installed with pip, and imported by other Python
programs.

This commit adopts the pyproject.toml spec for specifying project
metadata and setuptools for build system with a src-layout.

`wise` was used for the package top level name.  This is also the name
for the Python distribution.  This may cause issues later because the
distribution name is already taken on PyPI.

All Python modules were added into the wise package and __init__.py
files added as required.  Command line entry points were not moved
into the package yet.  The config.py file was moved to the wise.api
subackage because it only defines APIConfig.

Documentation for installation and testing were updated accordingly.