Remove unnecessary code for disabling thumbnail extraction for audio-only files
What this removed bit of code does is it calls `get_media_type(video_stream_info, audio_stream_info)` to determine the media type of the current file based on whether it has an audio stream and/or a video stream given by torchaudio.io.StreamReader. If the media type is a pure audio file, then we set `thumbnails = False` because we cannot extract thumbnails for audio-only files.

However, this bit of code is unnecessary because:
- When the AudioDataset class is initialised, we already set `thumbnails=False` (in `AudioDataset.__init__(...)`)
- In a25a3346 we are already using the output of `get_media_type(video_stream_info, audio_stream_info)` to determine which MediaDataset subclass (AudioDataset, VideoDataset, etc.) to use