Use MIME type rather than media codec to determine whether a media file is an image
Currently in the `get_media_type()` function in `src/dataloader/streamreader.py`, the media type is determined by checking if the media codec matches a list of known image codecs. However, when I tested this on a set of JPEG images, the codec returned by torchaudio.io.StreamReader was `mjpeg` which could also be a video codec. Hence, it might not be a good idea to use the codec to determine whether a media file is an image or video. Instead, it might be better to use the MIME type to do this.