Save actual file format (based on MIME type) instead of codec
Currently the `MediaMetadata.format` field is saved in the internal metadata database and used in `api/routes.py` to specify the MIME type when serving a media file. However, currently the codec (e.g. `h264` or `opus`) is saved in this field (instead of the actual file format) which means the MIME types returned in `api/routes.py` might be invalid.