FeatureExtractor.extract_image_features: return list of ndarrays.
So far, each image would return a single feature vector.  We will be
having models that return any number of feature vectors for each image
(effectively, a feature vector for each feature the model finds in
each image).

This commit simply changes extract_image_features to return a list so
that extract_image_features can continue to accept any number of
images as input and the output feature vectors can be assigned to the
correct image/frame timestamp.