Support for model specific vector metadata
FeatureExtractor get access to the database to add their own table and
record vector metadata.  For example, region coordinates, or detection
scores for individual vectors.

This commit only implements the framework to support it, it does not
add vector metadata to any of the existing feature extractors.  The
only behaviour change is that extract_image_features now returns an
object with vector and metadata attributes instead of a list of
vectors.