InsightFaceFeatureExtractor: feature vectors for face images with InsightFace
This adds support to search for faces with images, and supports any
number of faces per image.  Only the image modality is supported.

This only returns images with the given face and does not actually
displays, or even knows, where in the image are the faces.  That
requires adding support for vector metadata which is planned next.