Handle unsupported modalities during search on frontend.
The three modalities currently supported to search *with* are text,
image, and audio.  None of the models support all modalities and there
is no option to enable/disable modalities based on what the a model
supports.

This commit sets the methods related to unsupported modalities to None
(this is Python recommendation when parent uses NotImplementedError
and subclass wants to signal that it won't be implemented) so that
frontend knows the unsupported modalities and errors appropriately.