frontend: display bounding boxes if present in search results
This commit adds the ability to display bounding boxes if they
are present on ImageVector.  A new StillImageView component is
created to wrap the <img> with bounding boxes so it can be used
both on the main page and on the search results modal dialog.

Display of bounding boxes in videos is not yet supported.