StillImageView: replace Tooltip with Popover to add more components to it
At the moment, only the "distance" to the search is being displayed.
Next, we will want to display a button to "search similar" too and
in the future may display more details related to the vector.
Popover provides action elements like links and buttons which are
not available on Tooltip.