MediaInfo: replace empty type with parent MediaMetadata
MediaInfo is the type between MediaMetadata and the multiple *Info
types.  MediaInfo was previously the type that added search result
data to the media file metadata but search result is now based on the
vector and not media and all search result data is now part of the
VectorResult type.

Instead of just removing the MediaInfo type, this commit renames the
parent MediaMetadata type to MediaInfo.  This MediaMetadata type is
not used anywhere, it is not even exported.  It does this renaming to
make it obvious it is the parent of the other *Info types and also to
avoid confusion with the very similar, but not quite the same, class
in data_models.MediaMetadata.