Add FTS search wrapper class
This commit adds the skeleton required for metadata search. The class
takes in all the metadata tables (metadata-*) and the fts5 table if
present, and exposes an interface to run a fts5 match query across
all indexed columns.

This also takes care of ASR in a special way - handling highlights
that get split across segments, merging neighbouring segments, etc.

The search response is a VectorAndMediaMetadata, but the vector_id is set
to 'None'.

If there is an ASR match, the corresponding segments are returned as result segment.
If only media metadata matches the search, the first frame is returned as a result.