frontend (VectorInfo): move related_vectors property to Processed* type
The VectorInfo type in the frontend should mirror the one in the
backend.  The related_vectors is added by the frontend so should go
into the Processed* types.  It is not added during the first
"processing" steps, i.e., after the search fetch, it is only added
when the image details modal dialog is open but adding ProcessingLvl2*
types seems a bit too much.