Add video modal box overlay
Wrap the video player such that the box content can be shown on top of
a thumbnail, which is overlayed on top of the paused video, while allowing
user clicks to go to the video player outside the box. The user can still
interact with the box with the popover when paused and on timestamp of result.

Once played, the box vanishes and can only be displayed when user clicks on the
occurrence at the bottom.