create feature extractors using config passed from config.py
We can now pass config using the `feature_extractor_config` in config.py
This variable can be set with env var, wise_config.yaml and will be used
while creating feature extractor instances

It is handy when we instantiate feature extractors from project assets but want
to change the default behaviour. Example, to provide triton url for feature extractor
when serving the project.