LocalSearchService.filter_vectors: fix shot_scale_query
The variable `project_engine` is used both for metadata_query and
shot_scale_query but it is only defined/bound in the metadata_query
"block".  Move it one level up so it's available to shot_scale_query.

This fixes the case where only shot_scale_query is used triggering an
UnboundLocalError.