create-index.py: exit with error if fts_config is missing (revert 89b5577c)
When 89b5577c was applied the fix was drop the error exit status and
just log an error message instead.  The fix of omitting the metadata
key if the project has no metadata has now been applied (99e43324) so
we should go back to error properly if fts_config option is missing
when the project has metadata.  See discussion on issue #178.