test_feature_extractor.py: prevent albumentations from checking version
We disabled albumentations version check in feature.insightface and
relied on it when importing insightface.data again on
test_feature_extractor.  This worked because feature_extractor_factory
was importing all the feature extractors.  When that changed to have
the delay the import of feature extractors (see bd40b68d), the
disabling of albumentations version checker here also stopped working.

So just repeat it in test_feature_extractor to prevent a similar thing
from happening again.